requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
PyYAML>=5.4.0
pathlib2==2.3.7 